move editor-specific files into separate directory, make controllers directory
