+++ src/editor/contexts/GateReferenceContext.js
@@ -1,5 +1,6 @@
-import HiddenGraphContext from "calchart/contexts/HiddenContext";
 import Coordinate from "calchart/Coordinate";
+
+import HiddenGraphContext from "editor/contexts/HiddenContext";
 
 import HTMLBuilder from "utils/HTMLBuilder";
 import { round } from "utils/MathUtils";
